startReaderDrain: close the reader on exit
If one of our readers crash, we may end up blocking the running process
due to I/O buffering: the process tries to write to `stdout` but the
pipes are full and we're not reading them, so it hangs.

On our end, we're still waiting on `stderr` to close (and for the
process to exit), so we get a deadlock.

Note: right now, we *can* get such a crash fairly easily by writing
lines that are too long to `stdout` and crashing the reader with a
`bufio.Scanner: token too long` error.