startReaderDrain: close before reporting done
Calling `Wait()` on the process will cause the runtime to close our
reader pipes. Since we're deliberately closing the pipes ourselves, we
need to do so *before* calling `Wait()`.

`Wait()` is called once the wait group completes, so we can do this by
simply closing the pipe before reporting we're done.