Update cronexpr to reject out-of-order ranges
If an interval if out of order, it ends up emitting 0000-00-00 00:00:00
as the "next" time, which is undesirable. I patched cronexpr to reject
such intervals, since a workaround is to use e.g. `6,7` instead of
`6-7`.

Since the rejection reason might not be super obvious, this patch also
adds support for printing the errors from cronexpr. That requires
`-debug` for now, largely because we don't know exactly what error to
print (we try various lengths and only some of the errors might be
relevant).

Fixes #63