Vendor cronexpr in-tree
Rather than depend on a fork, let's incorporate cronexpr directly here.
Cronexpr won't ever receive updates at this point, so keeping it as a
fork so we could switch back to master is a bit pointless.

Incorporating it here into Supercronic gives us CI for free (which we
don't get in a fork), and it makes commits incorporating changes to
cronexpr clearer (since we can see the code changes instead of just a
dependency change).